requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.13.1
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
pytest==4.5.2
pytest-django==4.5.2
